Added phpcsCompliant setting
Enabling phpcsCompliant option, the extension will use snake case to create the getters and setters methods names; in addition it will use the correct number of spaces in comments and put the opening braces inline with the methods declarations.
If one or more of properties' names are not phpcs compliant an error is thrown and a message is shown to the user inviting him to disable phpcsCompliant options or to change properties names.